Fix warning: Default argument is mutable
Explanation of the warning in IntelliJ:

Reports a problem when a mutable value as a list or dictionary is detected in a default value for an argument. Default argument values are evaluated only once at function definition time, which means that modifying the default value of the argument will affect all subsequent calls of that function.
Example:
```
def func(s, cache={}):
     cache[s] = None
```
When the quick-fix is applied, the code changes to:
```
def func(s, cache=None):
     if cache is None:
              cache = {}
     cache[s] = None
``